code improvements and fixes
